# backend/openapi_milestone4.yaml
openapi: 3.0.3
info:
  title: PulseTrack HRMS API - Milestone 4
  version: 1.0.0
  description: |
    # PulseTrack HRMS - GenAI-Powered Human Resource Management System
    
    ## Project Information
    - **Team**: Team 11, IIT Madras
    - **Course**: Software Engineering (SEP 2025)
    - **Milestone**: 4 - API Endpoints Documentation
    
    ## Overview
    PulseTrack HRMS is a comprehensive, AI-powered Human Resource Management System designed to revolutionize 
    workforce management by making HR processes smarter, faster, and more human-centric.
    
    ## GenAI Integrations
    This system integrates multiple GenAI services to automate and enhance HR processes:
    
    ### 1. Google Gemini API
    - **Purpose**: Natural language processing, content generation, and intelligent analysis
    - **Used In**: Policy RAG, Resume Screening, Job Description Generation
    - **Model**: gemini-1.5-flash
    - **Features**: Text generation, embedding, multi-turn conversations
    
    ### 2. ChromaDB
    - **Purpose**: Vector database for semantic search
    - **Used In**: Policy RAG for document similarity search
    - **Features**: Persistent storage, similarity search, metadata filtering
    
    ### 3. Langchain
    - **Purpose**: RAG pipeline orchestration
    - **Used In**: Policy RAG question-answering system
    - **Features**: Document loading, text splitting, retrieval chains
    
    ### 4. PyPDF2
    - **Purpose**: PDF text extraction
...
    
...
    
...
    
...
    
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from /api/v1/auth/login endpoint"
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/leaves/me:
    get:
      tags:
        - Leave Management
      summary: Get My Leave Requests
      description: |
        **User Story #9: Leave Notifications**
        Receive timely notifications for leave approvals/rejections to stay informed.
        
        **Features:**
        - View all leave requests
        - See approval status
        - Filter by status, type, date range
        
        **Access:** All authenticated users
      operationId: getMyLeaves
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: leave_type
          in: query
          schema:
            type: string
            enum: [casual, sick, annual, maternity, paternity]
      responses:
        '200':
          description: Leave requests retrieved successfully
  /api/v1/leaves:
    post:
      tags:
        - Leave Management
      summary: Create Leave Request
      description: |
        **User Story #9: Leave Notifications**
        Submit a leave request. Notifications will be sent upon approval/rejection.
        
        **Access:** All authenticated users
      operationId: createLeaveRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leave_type:
                  type: string
                  enum: [casual, sick, annual, maternity, paternity]
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                reason:
                  type: string
              required:
                - leave_type
                - start_date
                - end_date
      responses:
        '201':
          description: Leave request created successfully
        '400':
          description: Bad Request - Insufficient leave balance
...